Move template tag utils to metabase-lib
